pyproject.toml
[project]
name = "marketplacebookstore"
version = "0.1.0"
description = "There is MarkteplaceBookStore"
authors = [
    {name = "Kate Nalivayko"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-tree-queries (>=0.19.0,<0.20.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cities-light (>=3.10.1,<4.0.0)",
    "django-smart-selects (>=1.7.2,<2.0.0)",
    "django-verify-email (>=3.0.3,<4.0.0)",
    "crispy-bulma (>=0.11.0,<0.12.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "pandas (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
